Split directory lookup logic to `wasm_support_fs` crate.
