+++ crate/asset_loading/Cargo.toml
@@ -21,4 +21,4 @@ serde_yaml = "0.8.11"
 strum = "0.18.0"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-web-sys = { version = "0.3.36", features = ["XmlHttpRequest"] }
+wasm_support_fs = { path = "../wasm_support_fs" }
