+++ crate/asset_loading/src/dir_traverse.rs
@@ -4,9 +4,9 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use log::{debug, error, warn};
+use log::{error, warn};
 #[cfg(target_arch = "wasm32")]
-use web_sys::XmlHttpRequest;
+use wasm_support_fs::DirAccess;
 
 /// Functions to make directory traversal code more ergonomic.
 #[derive(Debug)]
@@ -33,77 +33,7 @@ impl DirTraverse {
 
         #[cfg(target_arch = "wasm32")]
         {
-            Self::lookup_child_dirs(dir)
-        }
-    }
-
-    #[cfg(target_arch = "wasm32")]
-    fn lookup_child_dirs(dir: &Path) -> Vec<PathBuf> {
-        let dir_str = format!("{}", dir.display());
-        #[cfg(windows)]
-        let dir_str = dir_str.replace('\\', '/');
-
-        let xhr = XmlHttpRequest::new().expect("Failed to construct XmlHttpRequest");
-
-        // Synchronous GET request. Should only be run in web worker.
-        xhr.open_with_async("GET", dir_str.as_str(), false)
-            .expect("XmlHttpRequest open failed");
-
-        // We block here and wait for http fetch to complete
-        xhr.send().expect("XmlHttpRequest send failed");
-
-        // Status returns a result but according to javascript spec it should never return error.
-        // Returns 0 if request was not completed.
-        let status = xhr.status().expect("Failed to get XHR `status()`.");
-        match status {
-            200 => {
-                let response = xhr
-                    .response_text()
-                    .expect("Failed to get XHR `response_text()`.");
-
-                if let Some(response) = response {
-                    const SEARCH_TERM: &str = r#"style="font-weight: bold;" href=""#;
-                    let child_dirs = response
-                        .lines()
-                        .filter_map(|line| {
-                            line.find(SEARCH_TERM)
-                                .map(|find_index| find_index + SEARCH_TERM.len())
-                                .map(|path_start_index| &line[path_start_index..])
-                                .and_then(|path_to_end| {
-                                    path_to_end
-                                        .find('"')
-                                        .map(|path_end_index| &path_to_end[..path_end_index])
-                                })
-                        })
-                        .filter_map(|path_str| {
-                            if path_str.contains(".git") {
-                                None
-                            } else {
-                                #[cfg(windows)]
-                                let path_str = path_str.replace('/', '\\');
-
-                                Some(PathBuf::from(path_str))
-                            }
-                        })
-                        .collect::<Vec<PathBuf>>();
-
-                    warn!("Child directories for `{}`: {:?}", dir_str, child_dirs);
-
-                    child_dirs
-                } else {
-                    Vec::new()
-                }
-            }
-            404 => {
-                debug!("{} not found, returning empty directory list.", dir_str);
-                Vec::new()
-            }
-            _ => {
-                let msg = xhr.status_text().expect("Failed to get XHR `status_text`.");
-                error!("XmlHttpRequest failed with code {}. Error: {}", status, msg);
-
-                Vec::new()
-            }
+            DirAccess::child_dirs(dir)
         }
     }
