+++ crate/wasm_support_fs/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "wasm_support_fs"
+version = "0.19.0"
+authors = ["Azriel Hoh <mail@azriel.im>"]
+edition = "2018"
+
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+log = "0.4.8"
+web-sys = { version = "0.3.36", features = ["XmlHttpRequest"] }
